Split late stale terminal cleanup tests into a focused file.
Keeps stale-queued-work.test.ts under the 1k-line review bar without changing test behavior.
